Put full answer text in xapi statements
Adding this if/else has the script push the full text for the selected answer to xAPI statements rather than having the statement reference a number in the array of answers. This makes xAPI statements more readable and helps instructional technologists prepare visualizations and exports from learning record stores for instructors.